# openapi.json
{"openapi":"3.1.0","info":{"title":"KPI Service","version":"1.0"},"servers":[{"url":"https://kpi-portal-7474645160053590.aws.databricksapps.com","description":"Local dev"}],"paths":{"/dashboard/summary":{"get":{"summary":"Get Dashboard Summary","operationId":"get_dashboard_summary_dashboard_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardSummary"}}}}}}},"/kpis":{"get":{"summary":"List Kpis","operationId":"list_kpis_kpis_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/KPIResponse"},"title":"Response List Kpis Kpis Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/kpis/{kpi_id}":{"get":{"summary":"Get Kpi Detail","operationId":"get_kpi_detail_kpis__kpi_id__get","parameters":[{"name":"kpi_id","in":"path","required":true,"schema":{"type":"string","title":"Kpi Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KPIDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/kpis/{kpi_id}/favorite":{"post":{"summary":"Toggle Favorite","operationId":"toggle_favorite_kpis__kpi_id__favorite_post","parameters":[{"name":"kpi_id","in":"path","required":true,"schema":{"type":"string","title":"Kpi Id"}},{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drafts":{"post":{"summary":"Create Draft","operationId":"create_draft_drafts_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DraftCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Drafts","operationId":"list_drafts_drafts_get","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/drafts/{draft_id}/publish":{"post":{"summary":"Publish Kpi","operationId":"publish_kpi_drafts__draft_id__publish_post","parameters":[{"name":"draft_id","in":"path","required":true,"schema":{"type":"string","title":"Draft Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasource/tables":{"get":{"summary":"List Tables","description":"Get all tables + column definitions for selected schema","operationId":"list_tables_datasource_tables_get","parameters":[{"name":"schema","in":"query","required":false,"schema":{"type":"string","default":"gold_plus_datamart","title":"Schema"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TableListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasource/preview":{"get":{"summary":"Preview Table","description":"Get top 10 preview rows from selected table","operationId":"preview_table_datasource_preview_get","parameters":[{"name":"schema","in":"query","required":true,"schema":{"type":"string","title":"Schema"}},{"name":"table","in":"query","required":true,"schema":{"type":"string","title":"Table"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TablePreviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/query/optimize":{"post":{"summary":"Optimize Query","description":"Accept RAW SQL directly in body (no JSON escaping needed)","operationId":"optimize_query_query_optimize_post","requestBody":{"content":{"text/plain":{"schema":{"type":"string","title":"Sql"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/query/genie":{"post":{"summary":"Generate Sql With Genie","description":"Generate SQL using Databricks Genie","operationId":"generate_sql_with_genie_query_genie_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenieQueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/query/run":{"post":{"summary":"Execute Query","description":"Execute SQL and return preview rows","operationId":"execute_query_query_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteQueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/kpi/query-preparation":{"post":{"summary":"Query Preparation","operationId":"query_preparation_kpi_query_preparation_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryPreparationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/kpi/data-quality":{"post":{"summary":"Run Dq Checks","description":"Full enterprise Data Quality validation","operationId":"run_dq_checks_kpi_data_quality_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DQCheckRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/kpi/publish-final":{"post":{"summary":"Publish Final Kpi","description":"Insert KPI into kpi_master + kpi_values","operationId":"publish_final_kpi_kpi_publish_final_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublishKPIRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/{full_path}":{"get":{"summary":"Spa","operationId":"spa__full_path__get","parameters":[{"name":"full_path","in":"path","required":true,"schema":{"type":"string","title":"Full Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ColumnDef":{"properties":{"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type"}},"type":"object","required":["name","type"],"title":"ColumnDef"},"DQCheckRequest":{"properties":{"sql":{"type":"string","title":"Sql"},"metadata_signature":{"type":"string","title":"Metadata Signature"},"lineage_signature":{"type":"string","title":"Lineage Signature"}},"type":"object","required":["sql","metadata_signature","lineage_signature"],"title":"DQCheckRequest"},"DashboardChanges":{"properties":{"totalKPIs":{"type":"integer","title":"Totalkpis"},"activeKPIs":{"type":"integer","title":"Activekpis"},"categories":{"type":"integer","title":"Categories"},"thisMonth":{"type":"integer","title":"Thismonth"},"thisWeek":{"type":"integer","title":"Thisweek"}},"type":"object","required":["totalKPIs","activeKPIs","categories","thisMonth","thisWeek"],"title":"DashboardChanges"},"DashboardSummary":{"properties":{"totalKPIs":{"type":"integer","title":"Totalkpis"},"activeKPIs":{"type":"integer","title":"Activekpis"},"categories":{"type":"integer","title":"Categories"},"thisMonth":{"type":"integer","title":"Thismonth"},"thisWeek":{"type":"integer","title":"Thisweek"},"changes":{"$ref":"#/components/schemas/DashboardChanges"}},"type":"object","required":["totalKPIs","activeKPIs","categories","thisMonth","thisWeek","changes"],"title":"DashboardSummary"},"DownstreamUsage":{"properties":{"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type"},"frequency":{"type":"string","title":"Frequency"}},"type":"object","required":["name","type","frequency"],"title":"DownstreamUsage"},"DraftCreate":{"properties":{"user_id":{"type":"string","title":"User Id"},"kpi_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Kpi Name"},"data_source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Source"},"step_number":{"type":"integer","title":"Step Number","default":1}},"type":"object","required":["user_id"],"title":"DraftCreate"},"ExecuteQueryRequest":{"properties":{"sql":{"type":"string","title":"Sql"},"limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Limit","default":100}},"type":"object","required":["sql"],"title":"ExecuteQueryRequest"},"GenieQueryRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"table":{"type":"string","title":"Table"}},"type":"object","required":["prompt","table"],"title":"GenieQueryRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"KPIDetail":{"properties":{"kpi_id":{"type":"string","title":"Kpi Id"},"kpi_name":{"type":"string","title":"Kpi Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"business_formula":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Business Formula"},"sql_definition":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sql Definition"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"frequency":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Frequency"},"owner_team":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner Team"},"data_source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Source"},"business_unit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Business Unit"},"quality_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Quality Score"}},"type":"object","required":["kpi_id","kpi_name","description","business_formula","sql_definition","category","frequency","owner_team","data_source","business_unit","quality_score"],"title":"KPIDetail"},"KPIResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"status":{"type":"string","title":"Status"},"frequency":{"type":"string","title":"Frequency"},"owner":{"type":"string","title":"Owner"},"lastUpdated":{"type":"string","format":"date-time","title":"Lastupdated"},"sqlDefinition":{"type":"string","title":"Sqldefinition"},"qualityScore":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Qualityscore"},"linkedAssets":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Linkedassets"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"isFavorite":{"type":"boolean","title":"Isfavorite"},"definition":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Definition"},"businessFormula":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Businessformula"},"dataSource":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Datasource"},"businessUnit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Businessunit"},"complexity":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Complexity"},"nextUpdate":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Nextupdate"},"downstreamUsage":{"items":{"$ref":"#/components/schemas/DownstreamUsage"},"type":"array","title":"Downstreamusage","default":[]}},"type":"object","required":["id","name","status","frequency","owner","lastUpdated","sqlDefinition","qualityScore","linkedAssets","category","isFavorite","definition","businessFormula","dataSource","businessUnit","complexity","nextUpdate"],"title":"KPIResponse"},"PublishKPIRequest":{"properties":{"kpi_name":{"type":"string","title":"Kpi Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"business_formula":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Business Formula"},"sql":{"type":"string","title":"Sql"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"frequency":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Frequency"},"owner_team":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner Team"},"data_source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Data Source"},"business_unit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Business Unit"},"complexity":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Complexity"},"quality_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Quality Score"},"linked_assets":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Linked Assets"},"next_update":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Next Update"},"metadata_signature":{"type":"string","title":"Metadata Signature"},"semantic_signature":{"type":"string","title":"Semantic Signature"},"lineage_signature":{"type":"string","title":"Lineage Signature"}},"type":"object","required":["kpi_name","sql","metadata_signature","semantic_signature","lineage_signature"],"title":"PublishKPIRequest"},"QueryPreparationRequest":{"properties":{"sql":{"type":"string","title":"Sql"},"source_table":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Table"},"columns":{"items":{"type":"string"},"type":"array","title":"Columns","default":[]}},"type":"object","required":["sql"],"title":"QueryPreparationRequest"},"TableDef":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"schema":{"type":"string","title":"Schema"},"columns":{"items":{"$ref":"#/components/schemas/ColumnDef"},"type":"array","title":"Columns"}},"type":"object","required":["id","name","schema","columns"],"title":"TableDef"},"TableListResponse":{"properties":{"tables":{"items":{"$ref":"#/components/schemas/TableDef"},"type":"array","title":"Tables"}},"type":"object","required":["tables"],"title":"TableListResponse"},"TablePreviewResponse":{"properties":{"columns":{"items":{"type":"string"},"type":"array","title":"Columns"},"rows":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Rows"}},"type":"object","required":["columns","rows"],"title":"TablePreviewResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}