# src/api-client/models/DQCheckRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DQCheckRequest = {
    sql: string;
    metadata_signature: string;
    lineage_signature: string;
};