# src/api-client/models/DashboardChanges.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DashboardChanges = {
    totalKPIs: number;
    activeKPIs: number;
    categories: number;
    thisMonth: number;
    thisWeek: number;
};