# src/api-client/models/DashboardSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DashboardChanges } from './DashboardChanges';
export type DashboardSummary = {
    totalKPIs: number;
    activeKPIs: number;
    categories: number;
    thisMonth: number;
    thisWeek: number;
    changes: DashboardChanges;
};