# src/api-client/models/DownstreamUsage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DownstreamUsage = {
    name: string;
    type: string;
    frequency: string;
};