# src/api-client/models/DraftCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DraftCreate = {
    user_id: string;
    kpi_name?: (string | null);
    data_source?: (string | null);
    step_number?: number;
};