# src/api-client/models/ExecuteQueryRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExecuteQueryRequest = {
    sql: string;
    limit?: (number | null);
};