# src/api-client/models/GenieQueryRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GenieQueryRequest = {
    prompt: string;
    table: string;
};