# src/api-client/models/KPIDetail.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type KPIDetail = {
    kpi_id: string;
    kpi_name: string;
    description: (string | null);
    business_formula: (string | null);
    sql_definition: (string | null);
    category: (string | null);
    frequency: (string | null);
    owner_team: (string | null);
    data_source: (string | null);
    business_unit: (string | null);
    quality_score: (number | null);
};