# src/api-client/models/QueryPreparationRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type QueryPreparationRequest = {
    sql: string;
    source_table?: (string | null);
    columns?: Array<string>;
};