# src/api-client/models/TableDef.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ColumnDef } from './ColumnDef';
export type TableDef = {
    id: string;
    name: string;
    schema: string;
    columns: Array<ColumnDef>;
};