# src/api-client/models/TableListResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TableDef } from './TableDef';
export type TableListResponse = {
    tables: Array<TableDef>;
};