# src/api-client/models/TablePreviewResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TablePreviewResponse = {
    columns: Array<string>;
    rows: Array<Record<string, any>>;
};